requirements.txt
colorama
discord.py
PyYaml
